// packages/ctt/src/types/enums.ts
export enum TemplateUserFriendly {
  CalmsTypescriptApp = "Calm's Typescript App",
  CalmsTypescriptBase = "Calm's Typescript Base",
  CalmsTypescriptCdk = "Calm's Typescript CDK",
  CalmsTypescriptPackage = "Calm's Typescript Package",
}

export enum TypescriptExecutor {
  TsNode = 'ts-node',
  Tsx = 'tsx',
}